# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: TKA CBT Platform API
  version: 1.0.0
  license:
    name: Proprietary
    url: https://spdx.org/licenses/LicenseRef-Proprietary.html
  description: |
    Kontrak REST API untuk autentikasi, katalog dan pembayaran, CBT autosave,
    submit ujian, serta hasil analitik. Semua waktu menggunakan ISO 8601 UTC.
servers:
  - url: /
    description: Same host as the served API documentation
...
paths:
...
  /api/v1/packages:
    get:
      tags: [Packages]
      summary: List available tryout packages
      operationId: listPackages
      security: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Package list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/admin/payment-settings:
    get:
      tags: [Administration]
      summary: Get the active payment-provider configuration
      description: Owner-only. Secret values are never returned in plaintext.
      operationId: getPaymentSettings
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Masked payment-provider configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentSettingsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Administration]
      summary: Update the active payment-provider configuration
      description: |
        Owner-only. Omitted credentials and endpoint fields retain their
        current values. An explicitly empty redirect URL removes that redirect.
        Saved credentials are encrypted and activated without restarting the API.
      operationId: updatePaymentSettings
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentSettingsInput'
      responses:
        '200':
          description: Configuration saved and activated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentSettingsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token. The token JTI must match the active Redis session.
    WebhookSignature:
      type: apiKey
      in: header
      name: X-Payment-Signature
      description: HMAC signature supplied by the payment provider (optional).
    WebhookToken:
      type: apiKey
      in: header
      name: X-Callback-Token
      description: Webhook callback token that must match PAYMENT_WEBHOOK_SECRET.
  responses:
    BadRequest:
      description: Malformed request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing, invalid, expired, or replaced session token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: SESSION_INVALID
            message: Device Limit Reached / Session Invalid
            request_id: req_01JXYZ
    Forbidden:
      description: Authenticated user cannot access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Package:
      type: object
      required: [id, title, description, price, validity_days, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
          minimum: 0
        validity_days:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [active, inactive]
        created_at:
          type: string
          format: date-time
    PackageListResponse:
      type: object
      required: [data, page, per_page]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Package'
        page:
          type: integer
        per_page:
          type: integer
...
    PaymentSettingsInput:
      type: object
      additionalProperties: false
      properties:
        provider:
          type: string
          enum: [xendit]
          default: xendit
        secret_key:
          type: string
          format: password
          writeOnly: true
          description: Omit or send an empty string to retain the current secret key.
        webhook_token:
          type: string
          format: password
          writeOnly: true
          description: Omit or send an empty string to retain the current callback token; a new value must contain at least 32 characters.
        base_url:
          type: string
          description: Omit to retain the current URL; an empty string restores the environment/default URL.
        success_redirect_url:
          type: string
          description: Omit to retain the current URL; send an empty string to remove it.
        failure_redirect_url:
          type: string
          description: Omit to retain the current URL; send an empty string to remove it.
    PaymentSettings:
      type: object
      required: [provider, environment, secret_key_configured, webhook_token_configured, base_url, success_redirect_url, failure_redirect_url, updated_at]
      properties:
        provider:
          type: string
          enum: [xendit]
        environment:
          type: string
        secret_key_configured:
          type: boolean
        secret_key_masked:
          type: string
        webhook_token_configured:
          type: boolean
        webhook_token_masked:
          type: string
        base_url:
          type: string
          format: uri
        success_redirect_url:
          type: string
        failure_redirect_url:
          type: string
        updated_at:
          type: string
          format: date-time
    PaymentSettingsResponse:
      type: object
      required: [settings]
      properties:
        settings:
          $ref: '#/components/schemas/PaymentSettings'
...
    ErrorDetail:
      type: object
      required: [field, reason]
      properties:
        field:
          type: string
        reason:
          type: string
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Request validation failed
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
        request_id:
          type: string
          example: req_01JXYZ